+++ lmcache/v1/cache_controller/controllers/__init__.py
@@ -12,9 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from lmcache.experimental.cache_controller.controllers.kv_controller import \
+from lmcache.v1.cache_controller.controllers.kv_controller import \
     KVController
-from lmcache.experimental.cache_controller.controllers.registration_controller import \
+from lmcache.v1.cache_controller.controllers.registration_controller import \
     RegistrationController  # noqa: E501
 
 __all__ = [
