+++ lmcache/v1/cache_controller/controllers/registration_controller.py
@@ -17,10 +17,10 @@ from typing import Optional
 import zmq
 import zmq.asyncio
 
-from lmcache.experimental.cache_controller.message import (  # noqa: E501
+from lmcache.v1.cache_controller.message import (  # noqa: E501
     DeRegisterMsg, HealthMsg, HealthRetMsg, QueryInstMsg, QueryInstRetMsg,
     RegisterMsg)
-from lmcache.experimental.cache_controller.rpc_utils import (  # noqa: E501
+from lmcache.v1.cache_controller.rpc_utils import (  # noqa: E501
     close_zmq_socket, get_zmq_context, get_zmq_socket)
 from lmcache.logging import init_logger
